Removed the colors.py file and moved its content to __init__.py.
