+++ src/yesserpackageupdater/__init__.py
@@ -5,7 +5,13 @@ import asyncio
 import logging
 
 from logging import FileHandler
-from colors import Colors
+
+class Colors:
+    WHITE: str = '\u001b[37m'
+    RED: str = '\u001b[31m'
+    YELLOW: str = '\u001b[33m'
+    GREEN: str = '\u001b[32m'
+
 
 failed = ""
 outdated_count = 0
