--- src/yesserpackageupdater/colors.py
@@ -1,5 +1,0 @@
-class Colors:
-    WHITE: str = '\u001b[37m'
-    RED: str = '\u001b[31m'
-    YELLOW: str = '\u001b[33m'
-    GREEN: str = '\u001b[32m'
